pyproject.toml
[project]
name = "ai-employ"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "facebook-sdk>=3.1.0",
    "google-api-python-client>=2.192.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "openai>=2.28.0",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "tweepy>=4.16.0",
    "watchdog>=6.0.0",
    "pyyaml>=6.0",
]
